fix pre-commit: trailing whitespace, EOF, YAML, mypy, shellcheck
Whitespace/EOF (auto-fixed by hooks, now committed):
- Strip trailing whitespace from ~20 workflow/doc/script files
- Add missing EOF newlines across docs, examples, plugins, scripts

YAML:
- .github/ISSUE_TEMPLATE/social-campaign.yml was raw markdown; convert
  to valid GitHub issue template YAML (bare '@' caused check-yaml failure)

mypy:
- cli.py: add 'from typing import Any' + 'out: Any' declaration before
  the if/elif dispatch block; prevents dict[int, tuple[float, float]]
  being inferred as the type for all subsequent 'out' assignments
- my_recursive_ai.py: ROUTES holds mixed-signature lambdas; add
  # type: ignore[operator] on the call site (line 30)
- run_entropy_pump_harness.py: wrap return value in float() so mypy
  sees an explicit float instead of Any propagated via chess types

shellcheck:
- check_artifacts.sh: replace 'ls dist/*.whl | wc -l' with find (SC2012)
- check_evolution.sh: replace 'cat file | python' with 'python file' (SC2002)
- tools/release.sh: strip CRLF line endings (SC1017)

+++ scripts/run_entropy_pump_harness.py
@@ -38,7 +38,7 @@ def stock_eval(board: chess.Board) -> float:
         )
 
     ks = (shield(kW, chess.WHITE, +1) - shield(kB, chess.BLACK, -1)) * 0.5
-    return (m + mobility + ks) * 100.0  # cp
+    return float((m + mobility + ks) * 100.0)  # cp
 
 
 def eval_game(pgn_text: str, mid=(10, 30), tag="game"):
